refactor: inline `SingleLibraryItem` into `PublishLibrary` to reduce api surface area
